rustdoc: remove unused CSS selector `.methods > .item-info`
It was added with e08a84a0c18739417a50c3e46917ced5037244eb
(actually, it was called `.methods > .stability` at the time) and was
directly nested that way.

But with the switch to `<details>`, the code has changed drastically out from
under it, to the point where you have to go out of your way to actually get
it to render this way, and the result looks overly-tight and weird alongside
the normal version where this code is not reachable.