proc_macro/bridge: Eagerly pass TokenStream contents into the client
This change could improve performance, especially in cases like with the
cross-thread backend which have expensive RPC. It handles this by moving
more of the logic for TokenStream into the client, storing a
`Rc<Vec<TokenTree>>` directly, rather than converting to/from the
compiler representation over RPC for every change.

As this eagerly decomposes interpolated AST fragments, this ended up
changing the output of `TokenStream::to_string()`. To keep up
compatibility for now, formatting is still performed using RPC and
`rustc_ast_pretty`, however in the future we may look into decoupling
proc-macro formatting from `rustc_ast_pretty`, and doing it in the
client.

One side-effect of this was that the workaround for the
procedural-masquerade crate had to be updated to work with the new
approach, as we couldn't depend on the AST fragment printer to format
the input. To do this, a real `,` token is now inserted when passing the
enum to a proc-macro rather than depending on the pretty-printing
behaviour. From local tests it appears that the new output should be
compatible enough to work with older versions of procedural-masquerade.

The new conversion between the proc-macro and compiler representations
of a `TokenStream` is now handled by the `RpcContext` type provided by
the proc-macro server. This type is a distinct value which will be
fetched and stored by the server before commands are run, such that the
decoders can have access to relevant state. In the future, we may also
change the symbol handling code to work off of `RpcContext` rather than
being static methods on the `Server` trait.