Tell rustc about unused bits in Span.
This shrinks the size of `Option<Span>` and similar types, which shrinks
several AST and HIR nodes. The most important of these are `hir::Expr`
and `ast::Ty`.

This change was first attempted in #93747 where it had little effect.
But the improved niche-filling implemented by #94075 means this change
now has a much bigger effect.

Co-authored-by: Camille Gillot <gillot.camille@gmail.com>