Support `impl` in foreign modules for macro use
Allow `impl` blocks in foreign modules for more ergonomic use in macros.
`impl` blocks still have no meaning in foreign modules and will cause an
error if still present during lowering.

Also adds explicit tests for the previous similar foreign module macro
extensions to ensure that neither this nor those are accepted or crash
the compiler when *not* fed to a macro.