rustdoc: remove orphaned link on array bracket
This is 682889fb06591c4245422b73b005c5d8ae2d0cad, but for arrays instead.

For non-generics, this retains links to the array page, but instead of
trying to link it all, it only links the length part, which distinguishes
arrays from slices.

For generics, the entire thing becomes a link, just like slices.