Make the `config.src` handling for downloadable bootstrap a little more conservative
In particular, this supports build directories within an unrelated git repository.

As a side effect, it will fall back to the old logic when the source directory is being built from a tarball within an unrelated git repository.
However, that second case is unsupported and untested; we reserve the right to break it in the future.